+++ finerplan/forms.py
@@ -14,7 +14,7 @@ class AddTransactionForm(FlaskForm):
 	value = StringField("Value", validators=[DataRequired()])
 	transaction = RadioField("Type of Transaction",
                              choices=[('earnings', form_words['earnings']),
-                                      ('assets', form_words['assets']),
+                                      ('brokerage_transfers', form_words['brokerage_transfers']),
                                       ('expenses', form_words['expenses'])])
 
 	pay_method = RadioField("Payment Method",
@@ -22,22 +22,16 @@ class AddTransactionForm(FlaskForm):
 	                                  (form_words['credit'], form_words['credit']),
 	                                  (form_words['outsourced'], form_words['outsourced'])])
 	installments = IntegerField("Installments", default=1, validators=[DataRequired()])
-	category_0 = SelectField("Category", default="Mercado", choices = generate_categories('expenses'))
+	cat_expense = SelectField("Category", default="Mercado", choices = generate_categories('expenses'))
 	cat_earning = SelectField("Category", default="Mercado", choices = generate_categories('earnings'))
-	#cat_investment
-	#category_1 = SelectField("Category", choices = generate_categories()[1])
-	#category_2 = StringField("Category")
-#	pay_method = SelectField("Payment Method", choices = [(1, "Dinheiro"), (2, "Cartão de Crédito"), (3, "Terceiros")], validators=[DataRequired()], coerce=str)
-#	category_0 = SelectField("Category", choices = generate_categories()[0], validators=[DataRequired()])
-#	category_1 = SelectField("Category", choices = generate_categories()[1], validators=[DataRequired()])
-#	category_2 = SelectField("Category", choices = generate_categories()[2], validators=[DataRequired()])
+
 	submit = SubmitField("Add it")
 
 #class AddExpenseForm(AddTransactionForm):
 
 
 #class AddEarningForm(AddTransactionForm):
-#	category_0 = SelectField("Category", default="Mercado", choices = generate_categories()[0])
+#	cat_expense = SelectField("Category", default="Mercado", choices = generate_categories()[0])
 
 #class AddInvestmentForm(AddTransactionForm):